+++ server/src/modes/script/serviceHost.ts
@@ -8,7 +8,6 @@ import { LanguageModelCache } from '../languageModelCache';
 import { createUpdater, parseVueScript, isVue, isVueTemplate } from './preprocess';
 import { getFileFsPath, getFilePath } from '../../utils/paths';
 import * as bridge from './bridge';
-import * as chokidar from 'chokidar';
 
 // Patch typescript functions to insert `import Vue from 'vue'` and `new Vue` around export default.
 // NOTE: this is a global hack that all ts instances after is changed
@@ -43,7 +42,7 @@ const vueSys: ts.System = {
 
 if (ts.sys.realpath) {
   const realpath = ts.sys.realpath;
-  vueSys.realpath = function (path) {
+  vueSys.realpath = function(path) {
     if (isVueProject(path)) {
       return realpath(path.slice(0, -3)) + '.ts';
     }
@@ -78,19 +77,6 @@ export function getServiceHost(workspacePath: string, jsDocuments: LanguageModel
     ...parsedConfig.options
   };
   compilerOptions.allowNonTsExtensions = true;
-  const watcher = chokidar.watch(workspacePath, {
-    ignoreInitial: true,
-    ignored: defaultIgnorePatterns(workspacePath)
-  });
-
-  watcher
-    .on('change', filterNonScript(path => {
-      const ver = versions.get(path) || 0;
-      versions.set(path, ver + 1);
-    }))
-    .on('add', filterNonScript(path => {
-      files.push(path);
-    }));
 
   function updateCurrentTextDocument(doc: TextDocument) {
     const fileFsPath = getFileFsPath(doc.uri);
@@ -120,6 +106,12 @@ export function getServiceHost(workspacePath: string, jsDocuments: LanguageModel
       templateService: templateLanguageSerivice,
       scriptDoc: currentScriptDoc
     };
+  }
+
+  // External Documents: JS/TS, non Vue documents
+  function updateExternalDocument(filePath: string) {
+    const ver = versions.get(filePath) || 0;
+    versions.set(filePath, ver + 1);
   }
 
   function getScriptDocByFsPath(fsPath: string) {
@@ -239,11 +231,11 @@ export function getServiceHost(workspacePath: string, jsDocuments: LanguageModel
   const templateLanguageSerivice = ts.createLanguageService(templateHost, registry);
   return {
     updateCurrentTextDocument,
+    updateExternalDocument,
     getScriptDocByFsPath,
     dispose: () => {
-      watcher.close();
       jsLanguageService.dispose();
-    },
+    }
   };
 }
 
@@ -301,13 +293,4 @@ function getParsedConfig(workspacePath: string) {
     /*resolutionStack*/ undefined,
     [{ extension: 'vue', isMixedContent: true }]
   );
-}
-
-function filterNonScript(func: (path: string) => void) {
-  return (path: string) => {
-    if (!/(tsx?|vue|jsx?)$/.test(path)) {
-      return;
-    }
-    func(path);
-  };
 }