+++ server/src/modes/template/services/htmlCompletion.ts
@@ -17,7 +17,8 @@ export function doComplete(
   document: TextDocument,
   position: Position,
   htmlDocument: HTMLDocument,
-  tagProviders: IHTMLTagProvider[]
+  tagProviders: IHTMLTagProvider[],
+  emmetConfig: emmet.EmmetConfiguration
 ): CompletionList {
   const result: CompletionList = {
     isIncomplete: false,
@@ -26,7 +27,7 @@ export function doComplete(
 
   const offset = document.offsetAt(position);
   const node = htmlDocument.findNodeBefore(offset);
-  if (!node) {
+  if (!node || node.isInterpolation) {
     return result;
   }
   const text = document.getText();
@@ -274,16 +275,14 @@ export function doComplete(
           }
         }
         break;
+      case TokenType.Content:
+        if (offset <= scanner.getTokenEnd()) {
+          return emmet.doComplete(document, position, 'html', emmetConfig);
+        }
+        break;
       default:
         if (offset <= scanner.getTokenEnd()) {
-          return emmet.doComplete(document, position, 'html', {
-            useNewEmmet: true,
-            showExpandedAbbreviation: 'always',
-            showAbbreviationSuggestions: true,
-            syntaxProfiles: {},
-            variables: {},
-            preferences: {}
-          });
+          return result;
         }
         break;
     }