+++ server/src/modes/vue/index.ts
@@ -2,11 +2,19 @@ import { LanguageMode } from '../languageModes';
 import { doScaffoldComplete } from './scaffoldCompletion';
 
 export function getVueMode(): LanguageMode {
+  let config: any = {};
+
   return {
     getId() {
       return 'vue';
     },
+    configure(c) {
+      config = c;
+    },
     doComplete(document, position) {
+      if (!config.vetur.completion.useScaffoldSnippets) {
+        return { isIncomplete: false, items: [] };
+      }
       const offset = document.offsetAt(position);
       const text = document.getText().slice(0, offset);
       const needBracket = /<\w*$/.test(text);